refactor: ajusta estrutura do projeto
