+++ app/(tabs)/_layout.tsx
@@ -1,13 +1,13 @@
-import { Redirect, Tabs } from "expo-router";
-import React from "react";
-import { Platform } from "react-native";
-import { HapticTab } from "@/components/HapticTab";
-import { IconSymbol } from "@/components/ui/IconSymbol";
-import TabBarBackground from "@/components/ui/TabBarBackground";
-import { Colors } from "@/constants/Colors";
-import { useColorScheme } from "@/hooks/useColorScheme";
-import { useAuth } from "@/context/AuthContext";
-import { ExtratoProvider } from "../context/ExtratoContext";
+import { Redirect, Tabs } from 'expo-router';
+import React from 'react';
+import { Platform } from 'react-native';
+import { HapticTab } from '@/components/HapticTab';
+import { IconSymbol } from '@/components/ui/IconSymbol';
+import TabBarBackground from '@/components/ui/TabBarBackground';
+import { Colors } from '@/constants/Colors';
+import { useColorScheme } from '@/hooks/useColorScheme';
+import { useAuth } from '@/context/AuthContext';
+import { ExtratoProvider } from '@/app/context/ExtratoContext';
 
 const HomeIcon = ({ color }: { color: string }) => (
   <IconSymbol size={28} name="house.fill" color={color} />
@@ -31,37 +31,36 @@ export default function TabLayout() {
     <ExtratoProvider>
       <Tabs
         screenOptions={{
-          tabBarActiveTintColor: Colors[colorScheme ?? "light"].tint,
+          tabBarActiveTintColor: Colors[colorScheme ?? 'light'].tint,
           headerShown: false,
           tabBarButton: HapticTab,
           tabBarBackground: TabBarBackground,
           tabBarStyle: Platform.select({
             ios: {
               // Use a transparent background on iOS to show the blur effect
-              position: "absolute",
+              position: 'absolute',
             },
             default: {},
           }),
-        }}
-      >
+        }}>
         <Tabs.Screen
           name="index"
           options={{
-            title: "Home",
+            title: 'Home',
             tabBarIcon: HomeIcon,
           }}
         />
         <Tabs.Screen
-          name="transacoes"
+          name="transacoes/index"
           options={{
-            title: "Transações",
+            title: 'Transações',
             tabBarIcon: TransacoesIcon,
           }}
         />
         <Tabs.Screen
-          name="extrato"
+          name="extrato/index"
           options={{
-            title: "Extrato",
+            title: 'Extrato',
             tabBarIcon: ExtratoIcon,
           }}
         />
