+++ app/(tabs)/extrato/styles.ts
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  titleContainer: {
+    flexDirection: 'row',
+    gap: 8,
+    marginTop: 60,
+  },
+  emptyListMessage: {
+    textAlign: 'center',
+    marginTop: 20,
+  },
+  searchInput: {
+    height: 40,
+    borderColor: '#ccc',
+    borderWidth: 1,
+    borderRadius: 8,
+    paddingLeft: 8,
+    marginBottom: 16,
+    fontWeight: 'bold',
+  },
+});
