+++ app/(tabs)/transacoes/styles.ts
@@ -1,0 +1,16 @@
+import { StyleSheet, View } from 'react-native';
+
+export const styles = StyleSheet.create({
+  headerImage: {
+    color: '#808080',
+    bottom: -90,
+    left: -35,
+    position: 'absolute',
+  },
+  titleContainer: {
+    flexDirection: 'row',
+    gap: 8,
+    color: 'white',
+    marginTop: 60,
+  },
+});
