+++ app/context/ExtratoContext.tsx
@@ -1,15 +1,8 @@
-import React, {
-  createContext,
-  useContext,
-  useState,
-  useEffect,
-  ReactNode,
-} from "react";
-import { Alert } from "react-native";
-import { ItemPropsExtrato } from "@/components/utils/config";
-import extratoFirestore from "../services/extrato-firestore";
+import React, { createContext, useState, useEffect, ReactNode } from 'react';
+import { ItemPropsExtrato } from '@/components/utils/config';
+import extratoFirestore from '../services/extrato-firestore';
 
-interface ExtratoContextType {
+export interface ExtratoContextType {
   data: ItemPropsExtrato[];
   setData: React.Dispatch<React.SetStateAction<ItemPropsExtrato[]>>;
   fetchData: () => void;
@@ -18,19 +11,9 @@ interface ExtratoContextType {
   saldo: number;
 }
 
-const ExtratoContext = createContext<ExtratoContextType | undefined>(undefined);
+export const ExtratoContext = createContext<ExtratoContextType | undefined>(undefined);
 
-export const useExtrato = () => {
-  const context = useContext(ExtratoContext);
-  if (!context) {
-    throw new Error("useExtrato must be used within an ExtratoProvider");
-  }
-  return context;
-};
-
-export const ExtratoProvider: React.FC<{ children: ReactNode }> = ({
-  children,
-}) => {
+export const ExtratoProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
   const [data, setData] = useState<ItemPropsExtrato[]>([]);
   const [totalReceitas, setTotalReceitas] = useState(0);
   const [totalDespesas, setTotalDespesas] = useState(0);
@@ -42,15 +25,11 @@ export const ExtratoProvider: React.FC<{ children: ReactNode }> = ({
       setData(transactions);
 
       const receitas = transactions
-        .filter((item) => item.tipo === "deposit" || item.tipo === "reversal")
+        .filter(item => item.tipo === 'deposit' || item.tipo === 'reversal')
         .reduce((sum, item) => sum + item.valor, 0);
 
       const despesas = transactions
-        .filter((item) =>
-          ["payment", "withdraw", "transfer", "loan", "docted"].includes(
-            item.tipo
-          )
-        )
+        .filter(item => ['payment', 'withdraw', 'transfer', 'loan', 'docted'].includes(item.tipo))
         .reduce((sum, item) => sum + item.valor, 0);
 
       setTotalReceitas(receitas);
@@ -58,7 +37,7 @@ export const ExtratoProvider: React.FC<{ children: ReactNode }> = ({
       const despesasCorrigidas = Math.abs(despesas);
       setSaldo(receitas - despesasCorrigidas);
     } catch (error) {
-      console.error("Erro ao buscar transações:", error);
+      console.error('Erro ao buscar transações:', error);
     }
   };
 
@@ -68,8 +47,7 @@ export const ExtratoProvider: React.FC<{ children: ReactNode }> = ({
 
   return (
     <ExtratoContext.Provider
-      value={{ data, setData, fetchData, totalReceitas, totalDespesas, saldo }}
-    >
+      value={{ data, setData, fetchData, totalReceitas, totalDespesas, saldo }}>
       {children}
     </ExtratoContext.Provider>
   );
