+++ components/CardContainer/index.tsx
@@ -1,5 +1,6 @@
-import React, { FC } from "react";
-import { StyleSheet, View, Text, Image } from "react-native";
+import React, { FC } from 'react';
+import { View, Text, Image } from 'react-native';
+import { styles } from './styles';
 
 type CardContainerProps = {
   title?: string;
@@ -9,17 +10,17 @@ export const CardContainer: FC<CardContainerProps> = ({ title, children }) => {
   return (
     <View style={[styles.view]}>
       <Image
-        source={require("@/assets/images/textura.png")}
+        source={require('@/assets/images/textura.png')}
         style={[styles.imageRotate, { left: 0, top: 0 }]}
       />
       {title && <Text style={[styles.title]}>{title}</Text>}
       {children}
       <Image
-        source={require("@/assets/images/ilustration.png")}
+        source={require('@/assets/images/ilustration.png')}
         style={{ right: 0, bottom: 0, zIndex: 2 }}
       />
       <Image
-        source={require("@/assets/images/textura.png")}
+        source={require('@/assets/images/textura.png')}
         style={[styles.image, { right: 0, bottom: 0 }]}
       />
     </View>
@@ -27,35 +28,3 @@ export const CardContainer: FC<CardContainerProps> = ({ title, children }) => {
 };
 
 export default CardContainer;
-
-const styles = StyleSheet.create({
-  image: {
-    position: "absolute",
-    pointerEvents: "none",
-    zIndex: -1,
-  },
-  imageRotate: {
-    position: "absolute",
-    pointerEvents: "none",
-    zIndex: -1,
-    transform: [{ rotate: "180deg" }],
-  },
-  view: {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    gap: 32,
-    position: "relative",
-    backgroundColor: "#cbcbcb",
-    width: "100%",
-    padding: 32,
-    borderRadius: 8,
-    overflow: "hidden",
-  },
-  title: {
-    fontSize: 25,
-    fontWeight: "bold",
-    color: "#dee9ea",
-  },
-});
