+++ components/ExternalLink.tsx
@@ -1,4 +1,4 @@
-import { Link } from 'expo-router';
+import { ExternalPathString, Link } from 'expo-router';
 import { openBrowserAsync } from 'expo-web-browser';
 import { type ComponentProps } from 'react';
 import { Platform } from 'react-native';
@@ -10,8 +10,8 @@ export function ExternalLink({ href, ...rest }: Props) {
     <Link
       target="_blank"
       {...rest}
-      href={href}
-      onPress={async (event) => {
+      href={href as ExternalPathString}
+      onPress={async event => {
         if (Platform.OS !== 'web') {
           // Prevent the default behavior of linking to the default browser on native.
           event.preventDefault();
