+++ components/Loading/index.tsx
@@ -1,5 +1,6 @@
-import React from "react";
-import { View, StyleSheet, Animated, Easing } from "react-native";
+import React from 'react';
+import { View, Animated, Easing } from 'react-native';
+import { styles } from './styles';
 
 export const Loading = () => {
   const animationValue = new Animated.Value(0);
@@ -65,37 +66,5 @@ export const Loading = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    position: "relative",
-    width: 48,
-    height: 48,
-    flex: 0,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  loader: {
-    width: 12,
-    height: 12,
-    position: "relative",
-  },
-  dot: {
-    width: 12,
-    height: 12,
-    borderRadius: 6,
-    backgroundColor: "#FFF",
-    position: "absolute",
-  },
-  dot1: {
-    left: -18,
-  },
-  dot2: {
-    left: 0,
-  },
-  dot3: {
-    left: 18,
-  },
-});
 
 export default Loading;
